Made my own sorting algorithm.
Learning Python and decided to write a sorting function if .sort() didn't exist. It gets less efficient as duplicate numbers become more apparent. So probably best suited for unique id sorting. Also since it's doubling the variable size of the input list, probably don't want to use it on lists that may exceed storage limitations. Not bad for week two of learning.