Twitch Chat Bot Template
The class bot comes with two methods of incurring commands for the bot to do. First, there's the chat_trigger_text() function that MUST be called after the contents assignment. The second method is further down where you see @commands.command(). That gets called through the API if the user defined prefix is entered as the first non-space character.

Requires:
OAuth token given by Twitch and placed in line 8.
initial_channels (list) so the bot knows where to go. Line 8 as well.